Move LAP to replay_buffer module
